requirements.txt
Django==4.0.3
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
requests==2.27.1
celery==5.2.6
redis==4.2.2
